+++ strangeflix/strangeflix/settings.py
@@ -55,6 +55,10 @@ INSTALLED_APPS = [
     'transaction.apps.TransactionConfig',
     'provider.apps.ProviderConfig',
     'room.apps.RoomConfig',
+
+    # social accounts additional apps
+    'social_django', # add this
+
 ]
 
 MIDDLEWARE = [
@@ -117,6 +121,42 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
+# authentications backends for social account login to resolve the redirect url
+AUTHENTICATION_BACKENDS = [
+    'social_core.backends.linkedin.LinkedinOAuth2',
+    'social_core.backends.instagram.InstagramOAuth2',
+    'social_core.backends.facebook.FacebookOAuth2',
+    'social_core.backends.google.GoogleOAuth2',
+    'django.contrib.auth.backends.ModelBackend',
+]
+
+
+# redirecton urls for login with social accounts
+LOGIN_URL = 'login'
+LOGIN_REDIRECT_URL = 'social_redirect_url'
+LOGOUT_URL = 'logout'
+LOGOUT_REDIRECT_URL = 'home_page'
+
+
+# Facebook auth api configurations
+SOCIAL_AUTH_FACEBOOK_KEY = config('SOCIAL_AUTH_FACEBOOK_KEY')            # App ID
+SOCIAL_AUTH_FACEBOOK_SECRET = config('SOCIAL_AUTH_FACEBOOK_SECRET')      # App Secret
+SOCIAL_AUTH_FACEBOOK_SCOPE = ['email', 'user_link']                      # add this
+SOCIAL_AUTH_FACEBOOK_PROFILE_EXTRA_PARAMS = {                            # add this
+  'fields': 'id, name, email, picture.type(large), link'
+}
+SOCIAL_AUTH_FACEBOOK_EXTRA_DATA = [                 # add this
+    ('name', 'name'),
+    ('email', 'email'),
+    ('picture', 'picture'),
+    ('link', 'profile_url'),
+]
+
+
+# Google auth api configurations
+SOCIAL_AUTH_GOOGLE_OAUTH2_KEY = config('SOCIAL_AUTH_GOOGLE_OAUTH2_KEY')
+SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET = config('SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET')
+
 
 # Internationalization
 # https://docs.djangoproject.com/en/3.1/topics/i18n/
@@ -137,8 +177,8 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/3.1/howto/static-files/
 
 # login and logout redirection url
-LOGIN_REDIRECT_URL = 'home_page'
-LOGOUT_REDIRECT_URL = 'home_page'
+# LOGIN_REDIRECT_URL = 'home_page'
+# LOGOUT_REDIRECT_URL = 'home_page'
 
 # setting user auth model
 AUTH_USER_MODEL = config('AUTH_USER_MODEL')